Add and make use of the CommandGenerator class
The user has a list of commands that he can issue to the application like load/save a polygon object, undo, redo, clear, triangulate and so on.
The implementation of these commands lies in the KeyBindingsHandler class and initially they could only be invoked by pressing keys in order to generate key events. When the menu bar was added, we programmatically generated those key events in order to avoid code duplication. Now we'll start working on the toolbar soon, where we also need to generate key events by using almost the same code as the MenuBar. Thus the CommandGenerator was born.
The CommandGenerator is a singleton which can invoke the following commands by generating keypress/keyrelease events:
 - Load Polygon
 - Save Polygon
 - Exit
 - Undo
 - Redo
 - Clear
 - Add Points
 - Triangulate